Make time event IDs random and nonzero.
Which also nicely fixes a bug I was meaning to get around to: A double tap on check ammo key caused interface to completely lock up, because subsequent events with same ID confuse the event manager and the weapon is never brought up.

TIL: On Windows, PRNG sucks and frequently returns zero. That's why I chucked a 99999 in there, which means "return an integer with this upper bound and a lower bound of 1)