Move aliasing into on_game_start callback. Fixes MCM reset problem.
Calling into other modules while the script is loading is a bad idea. In this instance, calling `print_dbg` was indirectly forcing MCM to be executed early, before AC script was completely loaded. Which resulted in MCM not being able to call `on_mcm_load` in AC, thus, being unable to retrieve the config defaults.