Fix the "alife_clone_weapon" monkey business.
Attaching, detaching and swapping scopes (with or without ToggleScope) should not eat, change or create magazines anymore.

Also a small fix for dump_data so it doesn't crash and burn when thrown at a weapon without a magazine (magdata = false)

Note to future generations: Do not try to monkey patch _g (lowercase) because it is a local copy of the global environment created for each individual script. Modifying it will affect only _your_ script. If you really want to hack the global scope, you should refer to the _G (uppercase) table.