Remove magazine binder from weapons
Many addons may want to use script binding for weapons to implement different functionalities, so mags assigning binder to every weapon in the game just for first update is a big compatibility issue that is there for no reason. Instead we can create mag data right away... Not to mention that assigning magazine binder to weapons is somewhat counterintuitive.