SKS magazines moved from BaS patch to base mod
So, I saw that while the baseline mod does nothing for the SKS, the BaS patch does add perfectly serviceable magazines for the BaS SKS, which the vanilla SKS can't use them. This feels off, so I moved them to the base mod, and assigned them to the vanilla SKS as well.
These mags also had their model swapped with that of the AK alfa, so I corrected that as well.