BaS - Walther P99 extended magazine
This extended mag had no ties to any particular BaS element. I moved it to the basic files, with renamed section and string IDs to conform to the default one, and assigned it to the vanilla Walther as an improved mag.
The data for the old P99 BaS mag wasn't in use and I disabled it.