BaS - SIG 220 and old 9x39 extended magazines
Like with the extended P99 mag, these don't seem to have anything that impedes them being in the base files. The BaS 9x39 weapons themselves don't even have these assigned as improved mags.
We move them to the base files, and tweak them to conform to the current version.
We also disable the old default magazines, which are redundant.