BaS - Remington 700 mags
The 12 round mag: gets the WA2 world model like the base mag instead of the DVL mag model, which I feel was a bit too different; moved it to base files, and treat it as a extended one. Archangel and Lapua mags remain in the BaS patch, as they seem unique as per the weapon and kit's descriptions.
Assign the base 5 round mag as default to the BaS Remington 700 (as per the configs ammo_mag_size value).
Corrected the swapped strings assigned to name and description.
Base and BaS both add a basetype for the Remington 700; they're the same and BaS' isn't used. Disabled it.
Renamed all IDs, of base 5 round mag, and BaS 12 round, 20 round, and Lapua mags, to follow the same structure.